IKDREL3KK-295: printk: To remove spin lock for panic dump
Considering we need to do emergency_dump during printk, so we have to
remove the spin lock in buffer dump and buffer clear for apanic.
otherwise, system would lockup during emergency_dump.

Here we remove the spin lock in kmsg_dump_get_buffer_panic, which is
for apanic_mmc use only. Also, to remove the spinlock in buffer clear
interface, we add a new interface clear_log_buffer for apanic_mmc to
replace the exist interafce syslog_print_all.

The two interfaces would only be used when panic, that is, we disabled
interrupts and stopped other CPUs, so it is safe to remove the locks.

This is for apanic_mmc panic dump only.

Change-Id: Iff81866630cbdfde4f0076f94da0c0ad49edcdc2
Signed-off-by: Hong-Mei Li <a21834@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/623101
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Yi-Wei Zhao <gbjc64@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>