IKDREL3KK-1897: power: smb135x: Rerun APSD for SDP
The Slow Insertion detection of the SMB1359
does not appear to be slow enough.  DCP as still
getting detection as SDP 50% of the time. To fix
force a rerun of the APSD in the SMB1359 1 sec
after insertion and SDP detect.  This will only
run once so the User essentially has 1 sec to fully
insert the Device. This will cause a delay in the
Charger reporting by 1 sec as well but only in the
case that SDP is detected.

Change-Id: I95b298919b33f69fec9f096486cf1b58bbdeae3b
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.mot.com/635940
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-by: Ravikumar Vembu <raviv@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>