IKQCOMSD1-2572 arm/dt: msm8916: moto: disable pm8916 chg
We have our board configured for use with the smb1360, so we need to
disable the pm916_chg device. The pm8916_bms still needs to be left on,
but we need to use the 'disable' flag qcom added.

This eliminates a race condition on who probes first, and eliminates the
warning of duplicte battery_supply entries.

Change-Id: Ie2941439fd9cc48b3f47ee07788afb5998b43e33
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>