IKKRNBSP-2451 power: qpnp-cable-det: Add initial implementation
This driver is a subset of the QPNP linear/non-linear charger driver to
handle the USB related interrupts from the PMIC. This is to be used in
conjunction with an external charger-IC that does not have USB cable
detection functionality. This simply exposes a USB power supply and
handles the proper interrupts when a cable is inserted or removed.

Change-Id: Idc1f2f3b081983ffbe031ac61b75f8a4a4d663db
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-on: http://gerrit.mot.com/649863
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>