IKKRNBSP-2451 power: fan5404x: Re-start charging in handler
There is an issue being seen when plugging in a charger / usb cable
where we start charging but a few seconds later we get a pulse on the
STAT line and the TE, CE# and Voreg values have been reset to their
defaults. To workaround this, simply restart charging if we should be
charging.

Change-Id: I7c40e371f72e96e34e415b61cbf61d956981bfb3
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-on: http://gerrit.mot.com/650743
SLTApproved: Slta Waiver <sltawvr@motorola.com>