IKKRNBSP-2522 arm/dt: styx p1: Add max17042 Valrt threshold
Add a 'quirk' to the Styx P1s, so that the Valrt minimum threshold is
set to 3.5V. This is required since the P1 hardware will not boot when
the battery is below ~3.4V due to issue in the hardware connection of
the fan54046 and the pm8916. This workaround allows us to shutdown
before we reach that and still boot into dead-battery recovery.

Change-Id: Id11fb841aba745a6377388259c3bbaf64aee03aa
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-on: http://gerrit.mot.com/653623
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-by: Kenneth Kessler <kennykessler@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>