IKDREL3KK-8534: power: smb135x: Handle Interface Error
The Power Supply SYSFS interface was using functions
that just returned a value even if an error occured
This will change those functions so that it returns
an error and passes the value back through a pointer.

Note that invalid data could still propagate back to
Android if it is not handling the Errors on its end.
The Data has been sanitized to prevent a bad response
in case the error is not handled but the User could
still perceive the invalid values.
However internal calls to the functions from the
heartbeat work can now reschedule itself if it sees an
error rather then trying to operating using bad data.

Change-Id: Ic298c33a0106b50a0867d80c6290d03142ccb412
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.mot.com/662378
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Christopher Fries <cfries@motorola.com>
SLTApproved: Christopher Fries <cfries@motorola.com>
Reviewed-by: Vidyadhar Gurram <rnh786@motorola.com>
Reviewed-by: Lian-Wei Wang <lian-wei.wang@motorola.com>
Reviewed-by: Jiang-Li Yuan <jlyuan@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>