IKVPREL1L-277 arm/dt: 8916: Change INX display's VBP
The initial settings for the INX display had the top row of pixels missing.
Update the vertical back porch to fix this.

This will cause a difference in MIPI clock speed between the 2 display
sources for Surnia, but that is acceptable for display bring-up.

Change-Id: I7e0ffcd458d4f959a0f12087c00481f12365e0e7
Signed-off-by: Zac Crosby <zac@motorola.com>
Reviewed-on: http://gerrit.mot.com/664725
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Bang Nguyen <bangnguyen@motorola.com>
Reviewed-by: Stephen Rossbach <rossbach@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
SLTApproved: Christopher Fries <cfries@motorola.com>