IKDREL3KK-11007: power: smb135x: Stay Extended Low Cool
In the Extended Cool Range if the Battery Voltage is
detected as less than 4.0V the vFloat voltage is set
to 4.0V.  To prevent a flip flop with High Rate
Chargers be sure to stay in Low state if it is reached.
This will prevent the Charging/Discharging Toggle from
the fact that the Voltage will inflate once Charging is
re-enabled.

Change-Id: Ia533af77aba86d9176b8f76d0d8c128886accbe6
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.mot.com/666171
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Vidyadhar Gurram <rnh786@motorola.com>
Reviewed-by: Jiang-Li Yuan <jlyuan@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
SLTApproved: Christopher Fries <cfries@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>