IKDREL3KK-10246: power: smb135x: Prevent Recursive USB
The Force to set the USB online and present to false
when the Shutdown Voltage is tripped had a recursive
call.  This was causing rapid repeative calls to
the uevent and caused a system server crash.

The fix only sets the online and present to false
if the usb global is present and sets usb global to
false.

Change-Id: Ic5cb54d16c9b30701ae07c49fc51779c9f7e7a3a
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.mot.com/666232
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Jiang-Li Yuan <jlyuan@motorola.com>
Reviewed-by: Ravikumar Vembu <raviv@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
SLTApproved: Christopher Fries <cfries@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>