IKVPREL1L-567: power: fan5404x: Battery Temp Control
Allow the HEALTH parameter to be set externally so
that if the battery thermistor is hooked to a
different driver it can communicate the State.
Also keep track and monitor the Extended High
condition and control the battery charging accordingly.

Change-Id: I97eaee59f2f5a5d914c6dda97a90601250653cde
Signed-off-by: kumar venkatesh <kumarv@motorola.com>
Reviewed-on: http://gerrit.mot.com/670583
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
Reviewed-by: Vidyadhar Gurram <rnh786@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>