IKVPREL1L-747: power: fan5404x: Handle power-up with bootmode as factory.
Due to the absence of factory kill switch, the presence for USB accessory
and factory mode through utags is checked to trigger the powerdown. This
change handles the usecase to power-up the phone when the factory mode
through utags is set to TRUE with no USB accessory plugged in.

Change-Id: I1971458c324d6e6f2b7bdea52945cf57546a9c18
Signed-off-by: Vidyadhar Gurram <rnh786@motorola.com>
Reviewed-on: http://gerrit.mot.com/672866
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Jiang-Li Yuan <jlyuan@motorola.com>
Reviewed-by: Kenneth Kessler <kennykessler@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>