IKKRNBSP-2695 msm: mdss: Remove unnecessary locks and waits from ESD check
The wait for pingpong is only necessary if using the method of triggering
BTA to check the panel status.  Since we don't use this method, the wait
for pingpong can be removed.

Since we don't need to wait for pingpong, it is ok to remove the lock of
mdp5_data->ov_lock and allow new frames to be submitted during the status
check.  All locking necessary for sending commands for the status check
while continuing to send new frames is already handled.

Port of 9d03c7f.

Change-Id: Ic404bb5262a9b4281af3016b235c65453b30195f
Signed-off-by: Zac Crosby <zac@motorola.com>
Reviewed-on: http://gerrit.mot.com/678666
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Bang Nguyen <bangnguyen@motorola.com>
Reviewed-by: Shi-Yong Li <a22381@motorola.com>
Reviewed-by: Stephen Rossbach <rossbach@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>