IKVPREL1L-5111: sdhci-msm: increase hold-off time during hardware reset
Some platforms require more time for Vdd to drain when power-cycling
the card.  Increase both the initial hold-off time and the rate at
which the time increases to handle these.  This will be moved to the
device tree in a future revision of this patch so it can be tuned for
each platform.

Change-Id: I03d6bc62af5b85ee1187af9b768d25a4013588f5
Signed-off-by: Russ Knize <Russ.Knize@motorola.com>
Reviewed-on: http://gerrit.mot.com/708470
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Jason Hrycay <jason.hrycay@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>