msm: mdss: reset dirty flags for MDP3 LUT and CCS during ctrl off
LUT and CCS are programmed via asynchronous calls which won't be
in sync with display thread. If LUT has dirty flag set and ctrl off
happens, then on ctrl on, the LUT opmode will be programmed thereby
causing erroneous values of LUT to be applied to display. Hence
LUT should not have dirty flag set during ctrl off. For the case
of GC enabled, the LUT will be programmed appropriately with GC
values during resume and dirty flag will be set when needed. Hence
it will not be an issue to mark dirty as false here.

CCS currently gets checked and programmed for every commit cycle,
apart from the first programming via IOCTL call. However it is
good to have CCS dirty flag also set to false, in the off chance
that there are changes later that might overlook this same issue
scenario.

Change-Id: Ieda1f4de30a84f522dd76d0a4550b8c3c5b37c2f
Signed-off-by: Krishna Chaitanya Parimi <cparimi@codeaurora.org>