cluster_plug: disable by default
On ROMs not configured for use with ClusterPlug (eg. stock ROM), some components
(such as the PowerHAL and perfd) may misbehave. While I've tried to make it
tolerant of interference, it's best not to clash by default, and instead let
the users enable it themselves. On ROMs that properly support it (such as my
CyanogenMod port for lux), it will automatically be enabled by init scripts in
the ROM.